% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{derfinder-package}
\alias{derfinder-package}
\title{Generate HTML reports for results from fast differential expression analysis of RNA-seq data at base-pair resolution performed by derfinder}
\description{
Generate an HTML report using the results from derfinder (https://github.com/lcolladotor/derfinder). The HTML report itself is generated using knitrBoostrap (https://github.com/jimhester/knitrBootstrap) which uses knitr (http://yihui.name/knitr/) behind the scenes.
}
\author{
Leonardo Collado-Torres <lcollado@jhsph.edu>
}
\references{
Leonardo Collado-Torres, Alyssa Frazee, Andrew Jaffe and Jeffrey Leek (2013). derfinder: Fast differential expression analysis of RNA-seq data at base-pair resolution. https://github.com/lcolladotor/derfinder
}
\keyword{package}

