% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeReport.R
\name{edgeReport}
\alias{edgeReport}
\title{Generate a HTML/PDF report exploring edgeR results}
\usage{
edgeReport(dge, object, project = "", intgroup, colors = NULL,
  pAdjustMethod = "BH", alpha = 0.1, independentFiltering = FALSE, filter,
  theta, filterFun, nBest = 500, nBestFeatures = 20, customCode = NULL,
  outdir = "edgeRExploration", output = "edgeRExploration",
  browse = interactive(), device = "png", template = NULL,
  searchURL = "http://www.ncbi.nlm.nih.gov/gene/?term=", theme = NULL,
  digits = 2, ...)
}
\arguments{
\item{dge}{A \link[edgeR]{DGEList} object.}

\item{object}{A \link[edgeR:DGEList-class]{DGEExact} or 
\link[edgeR:DGELRT-class]{DGELRT} object that contains p-values stored in
\code{object$table$PValue}.}

\item{pAdjustMethod}{the method to use for adjusting p-values, see 
\link[stats]{p.adjust}. This argument will be passed to 
\link[DESeq2]{results}.}

\item{alpha}{the significance cutoff used for optimizing the independent 
filtering (by default 0.1). If the adjusted p-value cutoff (FDR) will be a 
value other than 0.1, alpha should be set to that value. This argument will
be passed to \link[DESeq2]{results}.}

\item{independentFiltering}{logical, whether independent filtering should be 
applied automatically. By default it's set to \code{FALSE} in contrast with
the default used in \link[DESeq2]{results} to match \code{edgeR}'s behavior.}

\item{theta}{the quantiles at which to assess the number of rejections from independent filtering. This argument is passed \link[DESeq2]{results}.}

\item{filterFun}{an optional custom function as described in 
\link[DESeq2]{results}.}

\item{the}{vector of filter statistics over which the independent filtering 
will be optimized. By default the logCPM will be used if
\code{independentFiltering} is set to \code{TRUE}. It can also be a length
1 character vector specifying one of the column names of \code{object$table}.}

\item{inheritParams}{DESeq2Report}
}
\value{
An HTML report with a basic exploration for the given set of edgeR
results.
}
\description{
This function generates a HTML report with exploratory data analysis plots
for edgeR results created. Other output formats are possible such as PDF 
reports but they lose the interactivity. Users can easily append
to the report by providing a R Markdown file to \code{customCode}, or can
customize the entire template by providing an R Markdown file to
\code{template}.
}
\examples{

## Create example data using DEFormats
library('DEFormats')
counts <- simulateRnaSeqData()
group <- rep(c("A", "B"), each = 3)

## Create DGEList object
library('edgeR')
dge <- DGEList(counts, group = group)

## Perform DE analysis with edgeR
design <- model.matrix( ~ group)
dge <- estimateDisp(dge, design)
fit <- glmFit(dge, design)
lrt <- glmLRT(fit, coef = 2)

## The output will be saved in the 'DESeq2Report-example' directory
dir.create('edgeReport-example', showWarnings = FALSE, recursive = TRUE)

## Generate the HTML report
report <- edgeReport(dge, lrt, project = 'edgeR-example', intgroup = 'group',
    outdir = 'edgeReport-example')

if(interactive()) {
    ## Browse the report
    browseURL(report)
}

\dontrun{
## Note that you can run the example using:
example('edgeReport', 'regionReport', ask=FALSE)
}


}
\author{
Leonardo Collado-Torres
}

