% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_check}
\alias{load_check}
\title{Attempt to load the namespace of one or more packages}
\usage{
load_check(pkg)
}
\arguments{
\item{pkg}{A character vector with the names of the packages to check.}
}
\description{
This function uses requireNamespace to try to load one or more packages.
If a package is missing, it will suggest how to install it via Bioconductor
before quitting.
}
\details{
Updated after feedback from Marcel Ramos at
\url{https://github.com/leekgroup/recount/issues/14}
}
\author{
Leonardo Collado-Torres
}
