% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_install}
\alias{load_install}
\title{Attempt to load the namespace of a package and install it if it's missing}
\usage{
load_install(pkg, quietly = TRUE)
}
\arguments{
\item{pkg}{A single character vector with the name of the package.}

\item{quietly}{Whether to run requireNamespace and biocLite quietly or not.}
}
\description{
This function uses requireNamespace to try to load a package. But if it's
misssing it will then install it via Bioconductor.
}
