% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/defunct.R
\name{regionReport-defunct}
\alias{derfinder_report}
\alias{regionReport-defunct}
\alias{render_report}
\title{Defunct functions in package \sQuote{regionReport}}
\usage{
derfinder_report()

render_report()
}
\description{
These functions are defunct and no longer available.
}
\details{
The following functions are defunct; use the replacements indicated below:
 \itemize{

   \item{plot_cluster: \code{\link{plotCluster}}}
   \item{render_report: \code{\link{render_report}}}

 }
}

