% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{regionReport-deprecated}
\alias{derfinder_report}
\alias{regionReport-deprecated}
\title{Deprecated functions in package \sQuote{regionReport}}
\usage{
derfinder_report()
}
\description{
These functions are provided for compatibility with older versions
of \sQuote{regionReport} only, and will be defunct at the next release.
}
\details{
The following functions are deprecated and will be made defunct; use
 the replacement indicated below:
 \itemize{

   \item{plot_cluster: \code{\link{plotCluster}}}
   \item{render_report: \code{\link{render_report}}}

 }
}

