% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/derfinderReport.R
\name{derfinderReport}
\alias{derfinderReport}
\alias{derfinder_report}
\title{Generate a HTML report exploring the basic results from derfinder}
\usage{
derfinderReport(prefix, outdir = "basicExploration",
  output = "basicExploration", project = prefix, browse = interactive(),
  nBestRegions = 100, makeBestClusters = TRUE, nBestClusters = 2,
  fullCov = NULL, hg19 = TRUE, p.ideos = NULL, txdb = NULL,
  device = "CairoPNG", ...)
}
\arguments{
\item{prefix}{The main data directory path where
\link[derfinder]{mergeResults} was run. It should be the same as
\code{mergeResults(prefix)}.}

\item{outdir}{The name of output directory relative to \code{prefix}.}

\item{output}{The name of output HTML file (without the html extension).}

\item{project}{The title of the project.}

\item{browse}{If \code{TRUE} the HTML report is opened in your browser once
it's completed.}

\item{nBestRegions}{The number of region plots to make, ordered by area.}

\item{makeBestClusters}{If \code{TRUE}, \link[derfinderPlot]{plotCluster} is
used on the \code{nBestClusters} regions by area. Note that these plots take
some time to make.}

\item{nBestClusters}{The number of region cluster plots to make by taking
the \code{nBestClusters} regions ranked by area of the cluster.}

\item{fullCov}{A list where each element is the result from
\link[derfinder]{loadCoverage} used with \code{cutoff=NULL}. Can be
generated using \link[derfinder]{fullCoverage}.}

\item{hg19}{If \code{TRUE} then the reference is assumed to be hg19 and
chromosome lengths as well as the default transcription database
(TxDb.Hsapiens.UCSC.hg19.knownGene) will be used.}

\item{p.ideos}{A list where each element is the result of
\link[ggbio]{plotIdeogram}. If it's \code{NULL} and \code{hg19=TRUE} then
they are created for the hg19 human reference.}

\item{txdb}{Specify the transcription database to use for making the plots
for the top regions by area. If \code{NULL} and \code{hg19=TRUE} then
TxDb.Hsapiens.UCSC.hg19.knownGene is used.}

\item{device}{The graphical device used when knitting. See more at
http://yihui.name/knitr/options (\code{dev} argument).}

\item{...}{Arguments passed to other methods and/or advanced arguments.}
}
\value{
An HTML report with a basic exploration of the derfinder results.
}
\description{
This function generates a HTML report exploring the basic results from
derfinder (https://github.com/lcolladotor/derfinder). The HTML report itself
is generated using knitrBoostrap which uses knitr (http://yihui.name/knitr/)
behind the scenes. It works best after using \link[derfinder]{mergeResults}.
}
\examples{
## Load derfinder
library('derfinder')

## The output will be saved in the 'derfinderReport-example' directory
dir.create('derfinderReport-example', showWarnings = FALSE, recursive = TRUE)

## For convenience, the derfinder output has been pre-computed
file.copy(system.file(file.path('extdata', 'chr21'), package='derfinder',
    mustWork=TRUE), 'derfinderReport-example', recursive = TRUE)

\dontrun{
## If you prefer, you can generate the output from derfinder
initialPath <- getwd()
setwd(file.path(initialPath, 'derfinderReport-example'))

## Collapse the coverage information
collapsedFull <- collapseFullCoverage(list(genomeData$coverage),
    verbose=TRUE)

## Calculate library size adjustments
sampleDepths <- sampleDepth(collapsedFull, probs=c(0.5), nonzero=TRUE,
    verbose=TRUE)

## Build the models
group <- genomeInfo$pop
adjustvars <- data.frame(genomeInfo$gender)
models <- makeModels(sampleDepths, testvars=group, adjustvars=adjustvars)

## Analyze chromosome 21
analyzeChr(chr='21', coverageInfo=genomeData, models=models,
    cutoffFstat=1, cutoffType='manual', seeds=20140330, groupInfo=group,
    mc.cores=1, writeOutput=TRUE, returnOutput=FALSE)

## Change the directory back to the original one
setwd(initialPath)
}

## Merge the results from the different chromosomes. In this case, there's
## only one: chr21
mergeResults(chrs = '21', prefix = 'derfinderReport-example',
    genomicState = genomicState$fullGenome)

## Load the options used for calculating the statistics
load(file.path('derfinderReport-example', 'chr21', 'optionsStats.Rdata'))

## Generate the HTML report
report <- derfinderReport(prefix='derfinderReport-example', browse=FALSE,
    nBestRegions=1, makeBestClusters=FALSE,
    fullCov=list('21'=genomeDataRaw$coverage), optionsStats=optionsStats)


if(interactive()) {
    ## Browse the report
    browseURL(report)
}

\dontrun{
## Note that you can run the example using:
example('derfinderReport', 'regionReport', ask=FALSE)
}
}
\author{
Leonardo Collado-Torres
}
\seealso{
\link[derfinder]{mergeResults}, \link[derfinder]{analyzeChr},
\link[derfinder]{fullCoverage}
}

