% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionReport-package.R
\docType{package}
\name{regionReport-package}
\alias{regionReport-package}
\title{Generate HTML or PDF reports for a set of regions or DESeq2/edgeR results.}
\description{
Generate an HTML reports to explore a set of regions such as the
results from annotation-agnostic expression analysis of RNA-seq data at 
base-pair resolution performed by derfinder. The package can also be used
to generate reports for DESeq2 or edgeR results.
The HTML and PDF reports are generated using rmarkdown 
(http://rmarkdown.rstudio.com/).
}
\author{
Leonardo Collado-Torres <lcollado@jhsph.edu>
}
\keyword{package}
