% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{regionReport-package}
\alias{regionReport-package}
\title{Generate HTML reports for a aset of regions.}
\description{
Generate an HTML reports to explore a set of regions such as the
results from annotation-agnostic expression analysis of RNA-seq data at
base-pair resolution performed by derfinder.
The HTML report itself is generated using knitrBootstrap
(https://github.com/jimhester/knitrBootstrap).
}
\author{
Leonardo Collado-Torres <lcollado@jhsph.edu>
}
\keyword{package}

